Use int64 dtype for bin_id column instead of uint32
Change to follow the pandas 3.0 default RangeIndex column's dtype.

+++ pygmt/src/grdhisteq.py
@@ -243,7 +243,7 @@ class grdhisteq:  # noqa: N801
                 dtype={
                     "start": np.float32,
                     "stop": np.float32,
-                    "bin_id": np.uint32,
+                    "bin_id": np.int64,
                 },
                 index_col="bin_id" if output_type == "pandas" else None,
             )
